Re-comment skull for Windows MATLAB
Skull transformation previously commented out by Dengyu intending
to de-bug error.  However, error was specific to MATLAB running on
Mac OS, causing error to be thrown when dbslocs run on the Windows
MATLAB server.  This commit reverts to original code so as to be
run on the Windows MATLAB server, as intended.